% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Print "FARS Accident Data Mapping State-wise"}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The function argument which stores a user-input State Number as an integer.}

\item{year}{The function argument which stores a year specified by the user (integer datatype).}

\item{filename}{Stores the name of file corresponding to the year entered as a string.}

\item{data}{Stores the information in file specified by "filename" as Dataframe Table.}

\item{data.sub}{Contains information filtered from the file based on the specified State Number as a list.}
}
\value{
This function returns a graphical plot of accidents occurring in the specified State during the year mentioned.
}
\description{
This is a simple function which reads a State Number and a Year,
and graphically plots the number of accidents that occurred within the specific geography and timeframe.
}
\note{
An error message will be displayed if an invalid State Number has been specified.
      The user may customize this message inside "" of the \code{stop} argument

If there is no data corresponding to a valid State Number, no graphs will be plotted.
      The user may customize the corresponding message shown, inside "" of the \code{message} argument

For any unavailable Latitude or Longitude values of that State, a "TRUE" or "FALSE" value is assigned accordingly.
}
