% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Print "Summarize Yearly Information on FARS Accidents in US"}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The function argument which stores a set of years entered by the user as integers.}

\item{dat_list}{Dataframe storing information on months for all valid years.}
}
\value{
This function returns the summarized information on accidents for all months corresponding to the valid years entered
        The information returned is a Dataframe Table type object.
}
\description{
This function reads a set of user-input years,
and returns each occurrence of accidents in each month for all those years, after reading from the corresponding files.
}
