% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Print "Make FILE NAME of Accident Data"}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The function argument which stores the year entered by the user, as an integer.}
}
\value{
This function prints the name of the file corresponding to the year entered as a string object.
}
\description{
This is a function that reads the year entered by a user during function call,
and prints the file name corresponding to that year.
}
\note{
C-type formatting used here ("sprintf" to convert "year" from integer to string while printing message).
}
